+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### 1.3.4 (2025-09-17)
+1. fix `morebuiltins.utils.gen_id` wrong length issue.
+
 ### 1.3.3 (2025-09-01)
 1. add `morebuiltins.logs.LogHelper.handle_crash` to log uncaught exceptions.
 2. fix `morebuiltins.logs.LogHelper.bind_handler` to avoid adding duplicate queue handlers.